Add LIS2MDL driver & implement in Imu component (#16)
* add lis2mdl driver and implement in imu component

* fix formatting
