+++ FprimeZephyrReference/Components/CMakeLists.txt
@@ -3,3 +3,4 @@
 add_fprime_subdirectory("${CMAKE_CURRENT_LIST_DIR}/FatalHandler")
 add_fprime_subdirectory("${CMAKE_CURRENT_LIST_DIR}/Watchdog")
 add_fprime_subdirectory("${CMAKE_CURRENT_LIST_DIR}/Imu/")
+add_fprime_subdirectory("${CMAKE_CURRENT_LIST_DIR}/Lis2mdlDriver/")
