+++ FprimeZephyrReference/Components/Imu/Imu.hpp
@@ -53,18 +53,12 @@ class Imu final : public ImuComponentBase {
     //! Get the angular velocity reading from the IMU
     Components::Imu_AngularVelocity get_angular_velocity();
 
-    //! Get the magnetic field reading from the IMU
-    Components::Imu_MagneticField get_magnetic_field();
-
     //! Get the temperature reading from the IMU
     F64 get_temperature();
 
     // ----------------------------------------------------------------------
     // Member variables
     // ----------------------------------------------------------------------
-
-    //! Zephyr device stores the initialized LIS2MDL sensor
-    const struct device* lis2mdl;
 
     //! Zephyr device stores the initialized LSM6DSO sensor
     const struct device* lsm6dso;
